new extended selection support for tables
Selection now considers the internal state of the JTable,
i.e. whether row- or column- or cell selection is turned on,
defaulting to row-selection.  Getting the selection returns the
.valueAt of the selected cells, setting the selection still
works by index.  Various combinations of nil, true, vectors and
integers are returned and accepted as arguments respectively.
(Needs documentation and tests.)